% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_link.R
\name{all_link}
\alias{all_link}
\title{Print all links of a given roadnet}
\usage{
all_link(roadnet)
}
\arguments{
\item{roadnet}{A weighted-adjacent-matrix of a roadnet. Every element represents the
length of link identified by 2 points. The rownames and colnames of roadnet matrix
should also be given, which is the point index/name.}
}
\value{
A charactor vector contains all links in the roadnet.
}
\description{
\code{all_link} receives a roadnet matrix and print all links with sep '-'.
}
\examples{
A = matrix(c(0,2,6,0,2,0,5,7,6,5,0,1,0,7,1,0),nrow=4)
rownames(A) <- 11:14
colnames(A) <- 11:14
all_link(A)

}
