% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/road_net.R
\docType{data}
\name{road_net}
\alias{road_net}
\title{Road net matrix for demo}
\format{The matrix's is 10*10. Number '0' means there's no connection or connect to itself.}
\usage{
road_net
}
\description{
A demo road net matrix contains 10 points
}
\keyword{datasets}
